add function for setting uniforms
