+++ src/components/animation/types.ts
@@ -5,3 +5,12 @@ export enum UniformTypes {
   CIRCLE,
   SHOCKWAVE,
 }
+
+export type Props = {
+  type: 'circle' | 'zigzag' | 'shockwave';
+  width: number;
+  height: number;
+  numVertices: number;
+  allowWierdMode?: boolean;
+  dynamicColor?: boolean;
+}
